+++ src/engine/people/cityworkers/Bureaucrat.java
@@ -39,7 +39,7 @@ public class Bureaucrat extends CityWorker implements Cloneable
         workplace = null;
     }
     @Override
-    public void doSkill(long time) {
+    public void doSkill(double time) {
         // TODO: 4/9/2016  make sure this ties into ui/ai
         //really this has nothing to do , in fact the args don't allow for any sort of management.
 		//so just charge the salarygiver for time
